% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_files}
\alias{add_files}
\alias{add_files<-}
\title{Add List of Files from Additional Table}
\usage{
add_files(refs) <- value
}
\arguments{
\item{refs}{A data frame including reference entries imported from BibTeX
databases by \code{\link{read_bib}}. At least a column called 'bibtexkey'
have to be included in this table.}

\item{value}{A data frame listing files with respective bibtexkey and
MIME-Type. In this table the columns 'bibtexkey', 'file', and 'mime' are
mandatory. The occurrence of all bibtexkey values in 'value' will be
cross-checked and evenctually cause an error, thus you may clean this
data frame before using it.}
}
\value{
The same data frame 'refs' with an inserted or updated column 'file'.
}
\description{
This function attempts to harmonize relational databases listing files in
a separated relation (data frame).
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
