% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2bibentry.R
\name{bib2bibentry}
\alias{bib2bibentry}
\alias{bib2bibentry.lib_df}
\alias{bib2bibentry,lib_df-method}
\title{Convert lib_df to bibentry}
\usage{
bib2bibentry(x, ...)

\method{bib2bibentry}{lib_df}(x, ...)
}
\arguments{
\item{x}{A \code{lib_df} object to be converted.}

\item{...}{Further arguments passed among methods (not yet in use).}
}
\value{
An \code{bibentry} object.
}
\description{
Conversion method for \code{lib_df} objects into \code{bibentry}.
}
\examples{
## Read installed electronic library
Bib <- read_bib(x = file.path(path.package("biblio"),
        "LuebertPliscoff.bib"))

## Convert the first five entries
bib2bibentry(Bib[1:5, ])
}
