% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_classes.R
\name{comp_df-class}
\alias{comp_df-class}
\title{Comparison between two data frames}
\description{
An S3 class for compared data frames. A list containing added, deleted
entries on the regarding a key column and cells that are modified.
}
