% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\alias{compare_df,data.frame,data.frame,character-method}
\alias{compare_df,lib_df,lib_df,missing-method}
\title{Compare data frames}
\usage{
compare_df(x, y, key, ...)

\S4method{compare_df}{data.frame,data.frame,character}(x, y, key, ...)

\S4method{compare_df}{lib_df,lib_df,missing}(x, y, key, ...)
}
\arguments{
\item{x}{The reference data frame.}

\item{y}{The updated data frame.}

\item{key}{A character value with the name of the variable used as primary
key in the tables.}

\item{...}{Further arguments passed among methods.}
}
\description{
Detect changes between two versions of a data frame.
}
