% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lib.R
\name{new_lib}
\alias{new_lib}
\title{Creates empty reference list}
\usage{
new_lib(cols)
}
\arguments{
\item{cols}{A character vector including the names for the columns in the new
reference list.}
}
\description{
A empty reference list will be created using a installed list of standard
column names for BibTeX databases. The columns can be customized by setting
an argument to the parameter 'cols'.
}
