% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{Read BibTeX Databases}
\usage{
read_bib(bib, journaltitle = TRUE, df = TRUE, ...)
}
\arguments{
\item{bib}{Path to BibTeX file.}

\item{journaltitle}{Logical value indicating whether empty values in field
'journaltitle' should be filled with respective values in 'journal'.}

\item{df}{Logical value indicating whether the output should be converted to
a data frame (the default) or just provided as matrix.}

\item{...}{Further arguments passed to \code{\link{readLines}}.}
}
\description{
Reading BibTeX databases and importing into R as a data frame. All the fields
will be inserted as character values.
}
