% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ris2df}
\alias{ris2df}
\title{Import a RIS file to a data frame}
\usage{
ris2df(x, ...)
}
\arguments{
\item{x}{A character value indicating the path to the RIS file.}

\item{...}{Further arguments (not yet in use).}
}
\value{
A \code{data.frame} object.
}
\description{
Since this package is focussing on reference lists in BibTeX, a translation
routine needs to be developed. For keeping it simple, the read function will
be kept as internal function producing a data frame.
}
\keyword{internal}
