% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synopsis}
\alias{synopsis}
\title{References by Lueber and Pliscoff (2018)}
\format{
An object of class \code{lib_df} (inherits from \code{data.frame}) with 1701 rows and 23 columns.
}
\source{
\doi{10.5281/zenodo.60800}
}
\usage{
synopsis
}
\description{
Example of an object formatted as \linkS4class{lib_df}. This library is published
with the references of the book
\bold{Bioclimatic and vegetational synopsis of Chile} by
\href{https://www.uchile.cl/publicaciones/141285/sinopsis-bioclimatica-y-vegetacional-de-chile}{Luebert and Pliscoff (2017)}.
}
\examples{
data(synopsis)

## Import from installed bibtex file
synopsis <- read_bib(x = file.path(path.package("biblio"),
  "LuebertPliscoff.bib"))

}
\keyword{datasets}
