% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_report.R
\name{update_pg}
\alias{update_pg}
\alias{update_report}
\alias{update_report,data.frame,data.frame-method}
\alias{update_report,PostgreSQLConnection,character-method}
\title{Compare Bibtex-files with Postgres databases and update}
\usage{
update_report(db, bib, ...)

\S4method{update_report}{data.frame,data.frame}(db, bib, print_only = TRUE, ...)

\S4method{update_report}{PostgreSQLConnection,character}(db, bib, db_args = list(), bib_args = list(), ...)
}
\arguments{
\item{db}{Either a data frame imported by \code{\link[=read_pg]{read_pg()}} or a connection
established by \code{\link[=dbConnect]{dbConnect()}} to a reference database.}

\item{bib}{Either a data frame imported by \code{\link[=read_bib]{read_bib()}} or a path to a bibtex
file. This data set represent an updated version of 'db'.}

\item{...}{Further arguments passed among methods.}

\item{print_only}{Logical value indicating whether the outcome will be only
printed in the console or stored as a list.}

\item{db_args}{List of named arguments passed to \code{\link[=read_pg]{read_pg()}}.}

\item{bib_args}{List of named arguments passed to \code{\link[=read_bib]{read_bib()}}.}
}
\description{
When changes done in the Bibtex duplicated file, they can be briefly
displayed before export.
}
