% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pg.R
\name{write_pg}
\alias{write_pg}
\title{Writing References from Data Frame to a Postgres Table}
\usage{
write_pg(conn, name, value, ...)
}
\arguments{
\item{conn}{A connection established with \code{\link{dbConnect}}.}

\item{name}{A character vector with the names of the schema and the table.}

\item{value}{A data frame including the data to be inserted in the created
table.}

\item{...}{Further arguments passed to \code{\link{dbWriteTable}}.}
}
\description{
This function implement the creation of a Postgres table and its population
with data imported by \code{\link{read_bib}}. To import data from a Postgres
table created by this function just use \code{\link{dbReadTable}}.
}
