% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{as}
\alias{as}
\alias{coerce,lib_df,data.frame-method}
\alias{coerce,lib_df,bibentry-method}
\alias{coerce,bibentry,lib_df-method}
\title{Coerce 'lib_df' objects}
\arguments{
\item{x}{An object to be coerced.}
}
\description{
Coercion 'lib_df' objects.
}
\examples{
## Read installed library
Refs <- read_bib(x = file.path(path.package("biblio"),
        "LuebertPliscoff.bib"))

# Convert lib_df to data frame
as(Refs[1:5, ], "data.frame")

# Convert lib_df to bibentry
as(Refs[1:5, ], "bibentry")
}
