% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{comp_df-class}
\alias{comp_df-class}
\title{Compared libraries}
\description{
An S3 class for compared data frames. A list containing added, deleted
entries on the regarding a key column and cells that are modified.
}
