% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_keys.R
\name{detect_keys}
\alias{detect_keys}
\alias{detect_keys.character}
\title{Detect bibtexkeys used in an r-markdown document}
\usage{
detect_keys(x, ...)

\method{detect_keys}{character}(x, ...)
}
\arguments{
\item{x}{A character vector to be scanned, for instance a file imported
either by \code{\link[=readLines]{readLines()}} or \code{\link[yamlme:read_rmd]{yamlme::read_rmd()}}.}

\item{...}{Further arguments passed among methods.}
}
\value{
A data frame with two columns, \code{bibtexkey} for the found keys and \code{line}
with the line number of the occurrence of the key in the document.
}
\description{
This function screens a character vector (usually an imported r-markdown
document) for the use of citations by bibtexkeys ('@bibtexkey'), retrieving
the detected key with its occurrence in the vector, assuming each element as
a line of the original document.

This function is based on \code{bbt_detect_citations()} from the package
\href{rbbt}{https://github.com/paleolimbot/rbbt}.
}
\examples{
## Read installed r-markdown document
my_documents <- readLines(file.path(path.package("biblio"), "document.Rmd"))

## Screen for citations
cited_refs <- detect_keys(my_documents)
cited_refs

}
