% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.lib_df}
\alias{print,lib_df-method}
\alias{print.comp_df}
\alias{print,comp_df-method}
\title{Print content of lib_df objects}
\usage{
\method{print}{lib_df}(x, ...)

\method{print}{comp_df}(x, ...)
}
\arguments{
\item{x}{An object of class 'lib_df'.}

\item{...}{Further arguments passed among methods.}
}
\value{
An invisible object, printed in the console.
}
\description{
A method for a brief overview on the content of a \linkS4class{lib_df} or a
\linkS4class{comp_df} object.
}
\examples{
print(synopsis)
}
\author{
Miguel Alvarez
}
