% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bib.R
\name{write_bib}
\alias{write_bib}
\alias{write_bib.lib_df}
\title{Write BibTeX Files}
\usage{
write_bib(x, ...)

\method{write_bib}{lib_df}(x, filename, encoding = "UTF-8", ...)
}
\arguments{
\item{x}{A data frame with bibliographic entries.}

\item{...}{Further arguments passed to \code{\link[=file]{file()}}.}

\item{filename}{A character value with the path and the name of the file to
be written.}

\item{encoding}{Character value with the encoding (passed to \code{\link[=file]{file()}}).}
}
\value{
A bibtex file.
}
\description{
BibTeX databases can be created from data frames, interacting with Postgres
databases.
}
